Better preserving state when using padding
Summary: When we are using ListView/GridView with topPadding
and we scroll a little first itme such as will not be hidden.
(Still on the top of the lit but a little scrolled). Before this
scroll will not be preserved after rotation and this header item
will be presented on the top of the list without scroll.

Change-Id: Iacaa78331d5c0551985e481b8b1d78bda7aa017d
Reviewed-on: https://review.appunite.com/7338
Tested-by: Jenkins <jenkins@appunite.com>
Reviewed-by: Jacek Marchwicki <jacek.marchwicki@gmail.com>